DOM Overlay integration to BM Tour.
Summary:
Initial iteration of DOM Overlays. Getting this out there as is, lots of details remain but overall looks good.

Big TODOs:
In-VR overlays need to be polished.
In-VR models are too small.
In-VR models cannot be interacted with, add support for controls and also arrow buttons to allow rotation.
More optimizations/details in/out of vr, e.g. preloading models{F68628916}
{F68628918}
{F68628917}

Reviewed By: mikearmstrong001

Differential Revision: D5433126

fbshipit-source-id: 917b5f5