Start of Compositor structure which holds all rendering
Summary:
Compositor handles all rendering of panels and entities. It may become possible in the future to break it into a shared base implementation, and subclasses that contain specific rendering logic (raw webgl, Three.js, etc). Each frame, the compositor is instructed to update, and then either draw itself to the main canvas, or to a VRDisplay. All of this information is handled externally  the Compositor has no sense of frame rate, or whether the user is in VR.

Future diffs will add in full surface support and the ability to mount 3D entities, as well as arbitrary scene management to support the traditional style of React VR layout.

This also introduces a new test page that starts to flesh out the ideal frame loop structure.

Reviewed By: mikearmstrong001

Differential Revision: D7120094

fbshipit-source-id: c09212ac7538c73597217700927e129cda0cef59