Move cursor into Compositor
Summary:
Moving GuiSys cursor into Compositor for new runtime.
If cursor visibility is hidden, it never appears.
If cursor visibility is auto, it appears when the runtime tells it to
If cursor visibility is visible, it always appears

All of these are contingent on the current raycaster producing a cursor.

Reviewed By: MartinSherburn

Differential Revision: D7497676

fbshipit-source-id: e9076443bfa88c70d6f4e34a69a7fa23818a5fd4