Support using a file as bridge file for WebWorker
Summary:
There's an issue with Chrome that, if a web worker is running with a blob url(which the default web worker is using), the network request from this web workder will not have "referer" in the header.
This will cause security issues if a React360 app is hosted on a high secured CDN if it checks the referer.
Here we support an option in React360, if you pass in the path to a bridge file, it will use the bridge file to create the web worker than using a blob url.

Reviewed By: MartinSherburn

Differential Revision: D14260799

fbshipit-source-id: c1c84cd56be3fa4f84a59d1196e127b07c00da4e