Provide an alternative way of placing hotspot in TourAppTempalte
Summary:
Provide an alternative way of placing hotspot in TourAppTempalte. This way we only use one big cylinder surface and place hotspot on it by mapping x direction translate to 3d rotation.
This has some limitation comparing to the recommented way but it's simpler to use for cases that doesn't exceed the limitation.
An option `useDynamicSurface` is provide in `index.html` to switch between two ways.

Reviewed By: MartinSherburn

Differential Revision: D14454992

fbshipit-source-id: a5c45411c900eaf8b00dab3adc6197d08745bb9f